pyproject.toml
[project]
name = "notifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
    "slack-sdk>=3.35.0",
    "sqlalchemy>=2.0.40",
    "watchdog>=6.0.0",
]
